requirements.txt
Flask
psycopg2-binary
Flask-SQLAlchemy
gevent
celery
redis
PyJWT
Flask-RESTful
requests